+++ Editor/EditorCommentWindow.cs
@@ -1,4 +1,6 @@
-﻿using Sirenix.OdinInspector.Editor;
+﻿using System;
+using Sirenix.OdinInspector.Editor;
+using Sirenix.Utilities.Editor;
 using UnityEditor;
 using UnityEditor.Experimental;
 using UnityEngine;
@@ -14,6 +16,7 @@ namespace OcUtility.Editor
             wnd.minSize = new Vector2(700, 600);
             wnd.position = new Rect(500, 500, 700, 600);
         }
+
         protected override OdinMenuTree BuildMenuTree()
         {
             var tree = new OdinMenuTree();
@@ -23,11 +26,44 @@ namespace OcUtility.Editor
             {
                 var path = AssetDatabase.GUIDToAssetPath(assetGUIDs[i]);
                 var asset = AssetDatabase.LoadAssetAtPath<EditorCommentAsset>(path);
+
+#if UNITY_2021_1_OR_NEWER
+                tree.Add(path.Replace(".asset","").Replace("Assets/", ""), asset, EditorGUIUtility.GetIconForObject(asset));
+#else
+                tree.Add(path.Replace(".asset","").Replace("Assets/", ""), asset, EditorGUIUtility.ObjectContent(asset, asset.GetType()).image);
+#endif
                 
-                tree.Add(path.Replace(".asset",""), asset, EditorGUIUtility.GetIconForObject(asset));
             }
 
             return tree;
+        }
+
+        protected override void OnBeginDrawEditors()
+        {
+            var selected = MenuTree.Selection.SelectedValue as EditorCommentAsset;
+            SirenixEditorGUI.BeginHorizontalToolbar();
+            if (SirenixEditorGUI.ToolbarButton("에셋 생성"))
+            {
+                var folderPath = selected == null ?
+                    "Assets/Editor Default Resources/Editor Comments":
+                    AssetDatabase.GetAssetPath(selected).Replace($"{selected.name}.asset", "");
+                var asset = EditorComment.CreateAsset(folderPath, "New Comment");
+                ForceMenuTreeRebuild();
+            }
+
+            if (selected == null) GUI.enabled = false;
+            var bColor = GUI.backgroundColor;
+            GUI.backgroundColor = Color.red;
+            if (SirenixEditorGUI.ToolbarButton("에셋 삭제"))
+            {
+                if(!EditorUtility.DisplayDialog("에셋 삭제", $"정말 다음 에셋을 삭제하시겠습니까?\n{selected.name}", "삭제", "취소")) return;
+                AssetDatabase.DeleteAsset(AssetDatabase.GetAssetPath(selected));
+                ForceMenuTreeRebuild();
+            }
+            GUI.backgroundColor = bColor;
+            GUI.enabled = true;
+            SirenixEditorGUI.EndHorizontalToolbar();
+            base.OnBeginDrawEditors();
         }
     }
 }